+++ cmake/modules/FindUFE.cmake
@@ -141,6 +141,15 @@ if (UFE_INCLUDE_DIR AND EXISTS "${UFE_INCLUDE_DIR}/ufe/sceneSegmentHandler.h")
     message(STATUS "Maya has UFE scene segment API")
 endif()
 
+set(UFE_HAS_DISPLAY_NAME FALSE CACHE INTERNAL "ufeHasDisplayName")
+if(UFE_INCLUDE_DIR AND EXISTS "${UFE_INCLUDE_DIR}/ufe/attribute.h")
+    file(STRINGS ${UFE_INCLUDE_DIR}/ufe/attribute.h UFE_HAS_API REGEX "displayName")
+    if(UFE_HAS_API)
+        set(UFE_HAS_DISPLAY_NAME TRUE CACHE INTERNAL "ufeHasDisplayName")
+        message(STATUS "Maya has UFE Attribute displayName")
+    endif()
+endif()
+
 set(UFE_TRIE_NODE_HAS_CHILDREN_COMPONENTS_ACCESSOR FALSE CACHE INTERNAL "ufeTrieNodeHasChildrenComponentsAccessor")
 if(UFE_INCLUDE_DIR AND EXISTS "${UFE_INCLUDE_DIR}/ufe/trie.h")
     file(STRINGS ${UFE_INCLUDE_DIR}/ufe/trie.h UFE_HAS_API REGEX "childrenComponents")
