+++ lib/mayaUsd/resources/ae/usdschemabase/__init__.py
@@ -1,5 +1,5 @@
 from .ae_template import AETemplate
-from .custom_enum_control import customEnumControlCreator
 import ufe
 if hasattr(ufe.Attributes, "getEnums"):
+    from .enum_custom_control import customEnumControlCreator
     AETemplate.prependControlCreator(customEnumControlCreator)
