+++ lib/mayaUsd/resources/ae/usdschemabase/ae_template.py
@@ -13,14 +13,17 @@
 # limitations under the License.
 #
 
-from .custom_image_control import customImageControlCreator
+from .image_custom_control import customImageControlCreator
+from .metadata_custom_control import MetaDataCustomControl
+from .connections_custom_control import connectionsCustomControlCreator
+from .array_custom_control import arrayCustomControlCreator
+from .attribute_custom_control import getNiceAttributeName
 
 import collections
 import fnmatch
 from functools import partial
 import re
 import ufe
-import usdUfe
 import maya.mel as mel
 import maya.cmds as cmds
 import mayaUsd.ufe as mayaUsdUfe
@@ -28,34 +31,9 @@ import mayaUsd.lib as mayaUsdLib
 import maya.internal.common.ufe_ae.template as ufeAeTemplate
 from mayaUsdLibRegisterStrings import getMayaUsdLibString
 
-try:
-    # This helper class was only added recently to Maya.
-    import maya.internal.ufeSupport.attributes as attributes
-    hasAEPopupMenu = 'AEPopupMenu' in dir(attributes)
-except:
-    hasAEPopupMenu = False
-
-from maya.common.ui import LayoutManager, ParentManager
-from maya.common.ui import setClipboardData
-from maya.OpenMaya import MGlobal
-
 # We manually import all the classes which have a 'GetSchemaAttributeNames'
 # method so we have access to it and the 'pythonClass' method.
 from pxr import Usd, UsdGeom, UsdLux, UsdRender, UsdRi, UsdShade, UsdSkel, UsdUI, UsdVol, Kind, Tf, Sdr, Sdf
-
-nameTxt = 'nameTxt'
-attrValueFld = 'attrValueFld'
-attrTypeFld = 'attrTypeFld'
-
-# Helper class to push/pop the Attribute Editor Template. This makes
-# sure that controls are aligned properly.
-class AEUITemplate:
-    def __enter__(self):
-        cmds.setUITemplate('attributeEditorTemplate', pst=True)
-        return self
-
-    def __exit__(self, mytype, value, tb):
-        cmds.setUITemplate(ppt=True)
 
 # Custom control, but does not have any UI. Instead we use
 # this control to be notified from UFE when any attribute has changed
@@ -108,293 +86,6 @@ class UfeConnectionChangedObserver(ufe.Observer):
         # Nothing needed here since we don't create any UI.
         pass
 
-class MetaDataCustomControl(object):
-    # Custom control for all prim metadata we want to display.
-    def __init__(self, item, prim, useNiceName):
-        # In Maya 2022.1 we need to hold onto the Ufe SceneItem to make
-        # sure it doesn't go stale. This is not needed in latest Maya.
-        mayaVer = '%s.%s' % (cmds.about(majorVersion=True), cmds.about(minorVersion=True))
-        self.item = item if mayaVer == '2022.1' else None
-        self.prim = prim
-        self.useNiceName = useNiceName
-
-        # There are four metadata that we always show: primPath, kind, active, instanceable
-        # We use a dictionary to store the various other metadata that this prim contains.
-        self.extraMetadata = dict()
-
-    def onCreate(self, *args):
-        # Metadata: PrimPath
-        # The prim path is for display purposes only - it is not editable, but we
-        # allow keyboard focus so you copy the value.
-        self.primPath = cmds.textFieldGrp(label='Prim Path', editable=False, enableKeyboardFocus=True)
-
-        # Metadata: Kind
-        # We add the known Kind types, in a certain order ("model hierarchy") and then any
-        # extra ones that were added by extending the kind registry.
-        # Note: we remove the "model" kind because in the USD docs it states, 
-        #       "No prim should have the exact kind "model".
-        allKinds = Kind.Registry.GetAllKinds()
-        allKinds.remove(Kind.Tokens.model)
-        knownKinds = [Kind.Tokens.group, Kind.Tokens.assembly, Kind.Tokens.component, Kind.Tokens.subcomponent]
-        temp1 = [ele for ele in allKinds if ele not in knownKinds]
-        knownKinds.extend(temp1)
-
-        # If this prim's kind is not registered, we need to manually
-        # add it to the list.
-        model = Usd.ModelAPI(self.prim)
-        primKind = model.GetKind()
-        if primKind not in knownKinds:
-            knownKinds.insert(0, primKind)
-        if '' not in knownKinds:
-            knownKinds.insert(0, '')    # Set metadata value to "" (or empty).
-
-        self.kind = cmds.optionMenuGrp(label='Kind',
-                                       cc=self._onKindChanged,
-                                       ann=getMayaUsdLibString('kKindMetadataAnn'))
-
-        for ele in knownKinds:
-            cmds.menuItem(label=ele)
-
-        # Metadata: Active
-        self.active = cmds.checkBoxGrp(label='Active',
-                                       ncb=1,
-                                       cc1=self._onActiveChanged,
-                                       ann=getMayaUsdLibString('kActiveMetadataAnn'))
-
-        # Metadata: Instanceable
-        self.instan = cmds.checkBoxGrp(label='Instanceable',
-                                       ncb=1,
-                                       cc1=self._onInstanceableChanged,
-                                       ann=getMayaUsdLibString('kInstanceableMetadataAnn'))
-
-        # Get all the other Metadata and remove the ones above, as well as a few
-        # we don't ever want to show.
-        allMetadata = self.prim.GetAllMetadata()
-        keysToDelete = ['kind', 'active', 'instanceable', 'typeName', 'documentation']
-        for key in keysToDelete:
-            allMetadata.pop(key, None)
-        if allMetadata:
-            cmds.separator(h=10, style='single', hr=True)
-
-            for k in allMetadata:
-                # All extra metadata is for display purposes only - it is not editable, but we
-                # allow keyboard focus so you copy the value.
-                mdLabel = mayaUsdLib.Util.prettifyName(k) if self.useNiceName else k
-                self.extraMetadata[k] = cmds.textFieldGrp(label=mdLabel, editable=False, enableKeyboardFocus=True)
-
-        # Update all metadata values.
-        self.refresh()
-
-    def onReplace(self, *args):
-        # Nothing needed here since USD data is not time varying. Normally this template
-        # is force rebuilt all the time, except in response to time change from Maya. In
-        # that case we don't need to update our controls since none will change.
-        pass
-
-    def refresh(self):
-        # PrimPath
-        cmds.textFieldGrp(self.primPath, edit=True, text=str(self.prim.GetPath()))
-
-        # Kind
-        model = Usd.ModelAPI(self.prim)
-        primKind = model.GetKind()
-        if not primKind:
-            # Special case to handle the empty string (for meta data value empty).
-            cmds.optionMenuGrp(self.kind, edit=True, select=1)
-        else:
-            cmds.optionMenuGrp(self.kind, edit=True, value=primKind)
-
-        # Active
-        cmds.checkBoxGrp(self.active, edit=True, value1=self.prim.IsActive())
-
-        # Instanceable
-        cmds.checkBoxGrp(self.instan, edit=True, value1=self.prim.IsInstanceable())
-
-        # All other metadata types
-        for k in self.extraMetadata:
-            v = self.prim.GetMetadata(k) if k != 'customData' else self.prim.GetCustomData()
-            cmds.textFieldGrp(self.extraMetadata[k], edit=True, text=str(v))
-
-    def _onKindChanged(self, value):
-        with mayaUsdLib.UsdUndoBlock():
-            model = Usd.ModelAPI(self.prim)
-            model.SetKind(value)
-
-    def _onActiveChanged(self, value):
-        with mayaUsdLib.UsdUndoBlock():
-            try:
-                usdUfe.ToggleActiveCommand(self.prim).execute()
-            except Exception as ex:
-                # Note: the command might not work because there is a stronger
-                #       opinion, so update the checkbox.
-                cmds.checkBoxGrp(self.active, edit=True, value1=self.prim.IsActive())
-                cmds.error(str(ex))
-
-    def _onInstanceableChanged(self, value):
-        with mayaUsdLib.UsdUndoBlock():
-            try:
-                usdUfe.ToggleInstanceableCommand(self.prim).execute()
-            except Exception as ex:
-                # Note: the command might not work because there is a stronger
-                #       opinion, so update the checkbox.
-                cmds.checkBoxGrp(self.instan, edit=True, value1=self.prim.IsInstanceable())
-                cmds.error(str(ex))
-
-# Custom control for all array attribute.
-class ArrayCustomControl(object):
-
-    if hasAEPopupMenu:
-        class ArrayAEPopup(attributes.AEPopupMenu):
-            '''Override the attribute AEPopupMenu so we can add extra menu items.
-            '''
-            def __init__(self, uiControl, ufeAttr, hasValue, values):
-                self.hasValue = hasValue
-                self.values = values
-                super(ArrayCustomControl.ArrayAEPopup, self).__init__(uiControl, ufeAttr)
-
-            def _copyAttributeValue(self):
-                setClipboardData(str(self.values))
-
-            def _printToScriptEditor(self):
-                MGlobal.displayInfo(str(self.values))
-
-            COPY_ACTION  = (getMayaUsdLibString('kMenuCopyValue'), _copyAttributeValue, [])
-            PRINT_ACTION = (getMayaUsdLibString('kMenuPrintValue'), _printToScriptEditor, [])
-
-            HAS_VALUE_MENU = [COPY_ACTION, PRINT_ACTION]
-
-            def _buildMenu(self, addItemCmd):
-                super(ArrayCustomControl.ArrayAEPopup, self)._buildMenu(addItemCmd)
-                if self.hasValue:
-                    cmds.menuItem(divider=True, parent=self.popupMenu)
-                    self._buildFromActions(self.HAS_VALUE_MENU, addItemCmd)
-
-    def __init__(self, ufeAttr, prim, attrName, useNiceName):
-        self.ufeAttr = ufeAttr
-        self.prim = prim
-        self.attrName = attrName
-        self.useNiceName = useNiceName
-        super(ArrayCustomControl, self).__init__()
-
-    def onCreate(self, *args):
-        attr = self.prim.GetAttribute(self.attrName)
-        typeName = attr.GetTypeName()
-        if typeName.isArray:
-            values = attr.Get()
-            hasValue = True if values and len(values) > 0 else False
-
-            # build the array type string
-            # We want something like int[size] or int[] if empty
-            typeNameStr = str(typeName.scalarType)
-            typeNameStr += ("[" + str(len(values)) + "]") if hasValue else "[]"
-
-            attrLabel = mayaUsdLib.Util.prettifyName(self.attrName) if self.useNiceName else self.attrName
-            singleWidgetWidth = mel.eval('global int $gAttributeEditorTemplateSingleWidgetWidth; $gAttributeEditorTemplateSingleWidgetWidth += 0')
-            with AEUITemplate():
-                # See comment in ConnectionsCustomControl below for why nc=5.
-                rl = cmds.rowLayout(nc=5, adj=3)
-                with LayoutManager(rl):
-                    cmds.text(nameTxt, al='right', label=attrLabel, annotation=attr.GetDocumentation())
-                    cmds.textField(attrTypeFld, editable=False, text=typeNameStr, font='obliqueLabelFont', width=singleWidgetWidth*1.5)
-
-                if hasAEPopupMenu:
-                    pMenu = self.ArrayAEPopup(rl, self.ufeAttr, hasValue, values)
-                    self.updateUi(self.ufeAttr, rl)
-                    self.attachCallbacks(self.ufeAttr, rl, None)
-                else:
-                    if hasValue:
-                        cmds.popupMenu()
-                        cmds.menuItem( label=getMayaUsdLibString('kMenuCopyValue'), command=lambda *args: setClipboardData(str(values)) )
-                        cmds.menuItem( label=getMayaUsdLibString('kMenuPrintValue'), command=lambda *args: MGlobal.displayInfo(str(values)) )
-
-        else:
-            errorMsgFormat = getMayaUsdLibString('kErrorAttributeMustBeArray')
-            errorMsg = cmds.format(errorMsgFormat, stringArg=(self.attrName))
-            cmds.error(errorMsg)
-
-    def onReplace(self, *args):
-        pass
-
-    # Only used when hasAEPopupMenu is True.
-    def updateUi(self, attr, uiControlName):
-        if not hasAEPopupMenu:
-            return
-
-        with ParentManager(uiControlName):
-            bgClr = attributes.getAttributeColorRGB(self.ufeAttr)
-            if bgClr:
-                isLocked = attributes.isAttributeLocked(self.ufeAttr)
-                cmds.textField(attrTypeFld, edit=True, backgroundColor=bgClr)
-
-    # Only used when hasAEPopupMenu is True.
-    def attachCallbacks(self, ufeAttr, uiControl, changedCommand):
-        if not hasAEPopupMenu:
-            return
-
-        # Create change callback for UFE locked/unlock synchronization.
-        cb = attributes.createChangeCb(self.updateUi, ufeAttr, uiControl)
-        cmds.textField(attrTypeFld, edit=True, parent=uiControl, changeCommand=cb)
-
-
-def showEditorForUSDPrim(usdPrimPathStr):
-    # Simple helper to open the AE on input prim.
-    mel.eval('evalDeferred "showEditor(\\\"%s\\\")"' % usdPrimPathStr)
-
-# Custom control for all attributes that have connections.
-class ConnectionsCustomControl(object):
-    def __init__(self, ufeItem, prim, attrName, useNiceName):
-        self.path = ufeItem.path()
-        self.prim = prim
-        self.attrName = attrName
-        self.useNiceName = useNiceName
-        super(ConnectionsCustomControl, self).__init__()
-
-    def onCreate(self, *args):
-        frontPath = self.path.popSegment()
-        attr = self.prim.GetAttribute(self.attrName)
-        attrLabel = self.attrName
-        if self.useNiceName:
-            attrLabel = mayaUsdLib.Util.prettifyName(self.attrName)
-            ufeItem = ufe.SceneItem(self.path)
-            if ufeItem:
-                try:
-                    ufeAttrS = ufe.Attributes.attributes(ufeItem)
-                    ufeAttr = ufeAttrS.attribute(self.attrName)
-                    if ufeAttr.hasMetadata("uiname"):
-                        attrLabel = str(ufeAttr.getMetadata("uiname"))
-                except:
-                    pass
-
-        attrType = attr.GetMetadata('typeName')
-
-        singleWidgetWidth = mel.eval('global int $gAttributeEditorTemplateSingleWidgetWidth; $gAttributeEditorTemplateSingleWidgetWidth += 0')
-        with AEUITemplate():
-            # Because of the way the Maya AE template is defined we use a 5 column setup, even
-            # though we only have two fields. We resize the main field and purposely set the
-            # adjustable column to 3 (one we don't have a field in). We want the textField to
-            # remain at a given width.
-            rl = cmds.rowLayout(nc=5, adj=3)
-            with LayoutManager(rl):
-                cmds.text(nameTxt, al='right', label=attrLabel, annotation=attr.GetDocumentation())
-                cmds.textField(attrTypeFld, editable=False, text=attrType, backgroundColor=[0.945, 0.945, 0.647], font='obliqueLabelFont', width=singleWidgetWidth*1.5)
-
-                # Add a menu item for each connection.
-                cmds.popupMenu()
-                for c in attr.GetConnections():
-                    parentPath = c.GetParentPath()
-                    primName = parentPath.MakeRelativePath(parentPath.GetParentPath())
-                    mLabel = '%s%s...' % (primName, c.elementString)
-
-                    usdSeg = ufe.PathSegment(str(c.GetPrimPath()), mayaUsdUfe.getUsdRunTimeId(), '/')
-                    newPath = (frontPath + usdSeg)
-                    newPathStr = ufe.PathString.string(newPath)
-                    cmds.menuItem(label=mLabel, command=lambda *args: showEditorForUSDPrim(newPathStr))
-
-    def onReplace(self, *args):
-        # We only display the attribute name and type. Neither of these are time
-        # varying, so we don't need to implement the replace.
-        pass
-
 
 class NoticeListener(object):
     # Inserted as a custom control, but does not have any UI. Instead we use
@@ -427,25 +118,10 @@ class NoticeListener(object):
                 if hasattr(ctrl, 'refresh'):
                     ctrl.refresh()
 
-def connectionsCustomControlCreator(aeTemplate, c):
-    if aeTemplate.attributeHasConnections(c):
-        return ConnectionsCustomControl(aeTemplate.item, aeTemplate.prim, c, aeTemplate.useNiceName)
-    else:
-        return None
-
-def arrayCustomControlCreator(aeTemplate, c):
-    # Note: UsdGeom.Tokens.xformOpOrder is a exception we want it to display normally.
-    if c == UsdGeom.Tokens.xformOpOrder:
-        return None
-
-    if aeTemplate.isArrayAttribute(c):
-        ufeAttr = aeTemplate.attrS.attribute(c)
-        return ArrayCustomControl(ufeAttr, aeTemplate.prim, c, aeTemplate.useNiceName)
-    else:
-        return None
-
 def defaultControlCreator(aeTemplate, c):
-    cmds.editorTemplate(addControl=[c])
+    ufeAttr = aeTemplate.attrS.attribute(c)
+    uiLabel = getNiceAttributeName(ufeAttr, c) if aeTemplate.useNiceName else c
+    cmds.editorTemplate(addControl=[c], label=uiLabel)
     return None
 
 class AEShaderLayout(object):
@@ -658,10 +334,14 @@ class AETemplate(object):
         for c in controls:
             if c not in self.suppressedAttrs:
                 for controlCreator in AETemplate._controlCreators:
-                    createdControl = controlCreator(self, c)
-                    if createdControl:
-                        self.defineCustom(createdControl, c)
-                        break
+                    try:
+                        createdControl = controlCreator(self, c)
+                        if createdControl:
+                            self.defineCustom(createdControl, c)
+                            break
+                    except Exception as ex:
+                        # Do not let one custom control failure affect others.
+                        print('Failed to create control %s: %s' % (c, ex))
                 self.addedAttrs.append(c)
 
     def suppress(self, control):
