+++ lib/mayaUsd/resources/ae/usdschemabase/enum_custom_control.py
@@ -13,26 +13,25 @@
 # limitations under the License.
 #
 
+from .attribute_custom_control import AttributeCustomControl
+
 import ufe
 import mayaUsd.ufe as mayaUsdUfe
 import mayaUsd.lib as mayaUsdLib
 import maya.cmds as cmds
 import maya.internal.ufeSupport.attributes as attributes
 
-class CustomEnumControl(object):
+class EnumCustomControl(AttributeCustomControl):
     def __init__(self, ufeAttr, ufeAttrType, prim, attrName, useNiceName):
-        self.ufeAttr = ufeAttr
+        super(EnumCustomControl, self).__init__(ufeAttr, attrName, useNiceName)
         self.prim = prim
         self.ufeAttrType = ufeAttrType
-        self.attrName = attrName
-        self.useNiceName = useNiceName
         ufeAttrs = ufe.Attributes.attributes(ufeAttr.sceneItem())
         self.enums = ufeAttrs.getEnums(ufeAttr.name)
-        super(CustomEnumControl, self).__init__()
 
     def onCreate(self, *args):
         # Create the control.
-        attrLabel = mayaUsdLib.Util.prettifyName(self.attrName) if self.useNiceName else self.attrName
+        attrLabel = self.getUILabel()
         self.uiControl = cmds.optionMenuGrp(label=attrLabel)
         attributes.AEPopupMenu(self.uiControl, self.ufeAttr)
 
@@ -48,7 +47,7 @@ class CustomEnumControl(object):
     def onReplace(self, *args):
         if len(self.enums) > 0:
             self.attachMenuItems()
-            attrLabel = mayaUsdLib.Util.prettifyName(self.attrName) if self.useNiceName else self.attrName
+            attrLabel = self.getUILabel()
             cmds.optionMenuGrp(self.uiControl, e=True, label=attrLabel)
         self.updateUi()
         self.attachCallbacks(self.updateEnumDataReader)
@@ -95,7 +94,7 @@ def customEnumControlCreator(aeTemplate, c):
     enums = ufeAttrs.getEnums(ufeAttr.name)
     # For now only integer enums are supported.
     if ufeAttrType == ufe.Attribute.kInt and len(enums) > 0:
-        return CustomEnumControl(ufeAttr, ufeAttrType, aeTemplate.prim, c, aeTemplate.useNiceName)
+        return EnumCustomControl(ufeAttr, ufeAttrType, aeTemplate.prim, c, aeTemplate.useNiceName)
     else:
         return None
