+++ lib/mayaUsd/resources/ae/usdschemabase/image_custom_control.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 #
 
+from .attribute_custom_control import AttributeCustomControl
+
 import mayaUsd.lib as mayaUsdLib
 from mayaUsdLibRegisterStrings import getMayaUsdLibString
 import mayaUsd_USDRootFileRelative as murel
@@ -26,29 +28,24 @@ from maya.common.ui import LayoutManager, ParentManager
 
 import ufe
 
-import os.path
-
 try:
     from PySide2 import QtCore
 except:
     from PySide6 import QtCore
 
 
-class ImageCustomControl(object):
+class ImageCustomControl(AttributeCustomControl):
     filenameField = "UIFilenameField"
 
     def __init__(self, ufeAttr, prim, attrName, useNiceName):
-        self.ufeAttr = ufeAttr
+        super(ImageCustomControl, self).__init__(ufeAttr, attrName, useNiceName)
         self.prim = prim
-        self.attrName = attrName
-        self.useNiceName = useNiceName
         self.controlName = None
-        super(ImageCustomControl, self).__init__()
 
     def onCreate(self, *args):
         cmds.setUITemplate("attributeEditorTemplate", pst=True)
 
-        attrUIName = mayaUsdLib.Util.prettifyName(self.attrName) if self.useNiceName else self.attrName
+        attrUIName = self.getUILabel()
         ufeAttr = self.ufeAttr
 
         createdControl = cmds.rowLayout(nc=3)
